requirements/dev.txt
-r base.txt
django-debug-toolbar==1.3.0
# Required by django-debug-toolbar
sqlparse==0.1.14

coverage==3.7.1
flake8==2.2.5
# Required by flake8
pyflakes==0.8.1
pep8==1.5.7
mccabe==0.3

PyYAML==3.10
Fabric==1.8.2
# Required by Fabric
paramiko==1.13.0
pycrypto==2.6.1
ecdsa==0.11

factory_boy==2.4.1
